TrailSafe/backend/pyproject.toml
[project]
name = "trailsafe-backend"
version = "0.1.0"
description = "TrailSafe cloud LLM proxy with provider-agnostic abstraction"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "openai>=1.55",
    "anthropic>=0.40",
    "google-genai>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
